fix problem : loading elements paggination

+++ src/app/back/claim/all/all.claim.back.ts
@@ -18,7 +18,7 @@ export class AllClaimBackComponent {
   allClaims: Claim[];
   tableChecked = new Array;
   claimToEdit: Claim;
-  config: any;
+  config: {} = {itemsPerPage: 0, currentPage: 0, totalItem: 0};
   collection = {count: 10, data: []};
 
   constructor(private claimService: ClaimService,
@@ -39,17 +39,6 @@ export class AllClaimBackComponent {
   edit(c: Claim) {
     const modalRef = this.modalService.open(EditClaimBackComponent);
     modalRef.componentInstance.c = c;
-  }
-
-  delete(c: Claim) {
-    this.claimService.deleteClaim(c).subscribe(
-      response => {
-        this.fetchData();
-      },
-      error => {
-        console.log(error);
-      }
-    );
   }
 
   fetchData() {
@@ -82,7 +71,7 @@ export class AllClaimBackComponent {
     );
   }
 
-  clickMethod(c: Claim) {
+  deleteClaim(c: Claim) {
     if (confirm('Are you sure to delete this claim')) {
       this.claimService.deleteClaim(c).subscribe(
         response => {
@@ -104,19 +93,21 @@ export class AllClaimBackComponent {
   }
 
   deleteSelectedClaims() {
-    console.log(this.tableChecked);
-    this.tableChecked.forEach(x => {
-      this.claimService.deleteClaim(x).subscribe(
-        response => {
-          console.log(this.tableChecked.indexOf(x));
-          this.allClaims.splice(this.tableChecked.indexOf(x), 1);
-          this.tableChecked.splice(this.tableChecked.indexOf(x), 1);
-        },
-        error => {
-          console.log(error);
-        }
-      );
-    });
+    if (confirm('Are you sure to delete those claims !!! ')) {
+      console.log(this.tableChecked);
+      this.tableChecked.forEach(x => {
+        this.claimService.deleteClaim(x).subscribe(
+          response => {
+            console.log(this.tableChecked.indexOf(x));
+            this.allClaims.splice(this.tableChecked.indexOf(x), 1);
+            this.tableChecked.splice(this.tableChecked.indexOf(x), 1);
+          },
+          error => {
+            console.log(error);
+          }
+        );
+      });
+    }
   }
 
   pageChanged(event) {
